requirements.txt
openai>=1.57.0,<2.0
httpx>=0.27.0,<0.29
flask==3.0.3
python-dotenv==1.0.1
pydantic==2.9.2
werkzeug==3.0.4